fix: modularize history dialogs and localize transaction strings
