+++ app/src/main/java/com/serranoie/app/minus/presentation/ui/history/edit/dialogs/EditDatePickerDialog.kt
@@ -1,4 +1,4 @@
-package com.serranoie.app.minus.presentation.ui.history.edit
+package com.serranoie.app.minus.presentation.ui.history.edit.dialogs
 
 import androidx.compose.material3.DatePicker
 import androidx.compose.material3.DatePickerDialog
@@ -8,13 +8,12 @@ import androidx.compose.material3.TextButton
 import androidx.compose.material3.rememberDatePickerState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import com.serranoie.app.minus.R
+import com.serranoie.app.minus.presentation.ui.theme.MinusTheme
 import java.time.Instant
 import java.time.LocalDate
-import java.time.ZoneId
 import java.time.ZoneOffset
-import androidx.compose.ui.tooling.preview.Preview
-import com.serranoie.app.minus.presentation.ui.theme.MinusTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
