+++ app/src/main/java/com/serranoie/app/minus/presentation/ui/history/edit/dialogs/EditTimePickerDialog.kt
@@ -1,5 +1,6 @@
-package com.serranoie.app.minus.presentation.ui.history.edit
+package com.serranoie.app.minus.presentation.ui.history.edit.dialogs
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -55,7 +56,7 @@ internal fun EditTimePickerDialog(
                     modifier = Modifier
                         .fillMaxWidth()
                         .padding(top = 16.dp),
-                    horizontalArrangement = androidx.compose.foundation.layout.Arrangement.End
+                    horizontalArrangement = Arrangement.End
                 ) {
                     TextButton(onClick = onDismiss) {
                         Text(stringResource(R.string.cancel))
